# eventbooking-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Event Booking Mini App",
    "description": "Mini-app dat ve su kien tich hop Super App. Cac endpoint co OpenAPI metadata de Agent co the tim kiem, uoc tinh gia, dat ve va huy ve.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://eventbooking-i19e.onrender.com"
    }
  ],
  "paths": {
...
    "/events": {
      "get": {
        "summary": "Tìm sự kiện phù hợp để người dùng xem và chọn vé",
        "description": "Dùng endpoint này khi người dùng muốn tìm sự kiện theo thành phố, từ khóa, danh mục, ngày hoặc ngân sách.",
        "operationId": "search_events",
        "parameters": [
          {
            "name": "city",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Thành phố tổ chức sự kiện (VD: Hà Nội, TP. Hồ Chí Minh, Đà Nẵng)",
              "title": "City"
            },
            "description": "Thành phố tổ chức sự kiện (VD: Hà Nội, TP. Hồ Chí Minh, Đà Nẵng)"
          },
          {
            "name": "keyword",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Từ khóa tên hoặc nội dung sự kiện",
              "title": "Keyword"
            },
            "description": "Từ khóa tên hoặc nội dung sự kiện"
          },
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Danh mục sự kiện. Giá trị gợi ý: conference | music | expo",
              "title": "Category"
            },
            "description": "Danh mục sự kiện. Giá trị gợi ý: conference | music | expo"
          },
          {
            "name": "start_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Ngày bắt đầu tìm kiếm sự kiện",
              "title": "Start Date"
            },
            "description": "Ngày bắt đầu tìm kiếm sự kiện"
          },
          {
            "name": "max_price",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "description": "Giá vé tối đa theo VNĐ",
              "title": "Max Price"
            },
            "description": "Giá vé tối đa theo VNĐ"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Search Events"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "x-superapp": {
          "capability": "event.search",
          "sideEffect": "read",
          "riskLevel": "low",
          "requiresConfirmation": false,
          "idempotency": "none"
        },
        "x-risk-level": "low",
        "x-side-effect-type": "read",
        "x-requires-hitl": false,
        "x-idempotency-required": false,
        "x-retry-policy": "safe_retry",
        "x-deep-link-template": "https://eventbooking-i19e.onrender.com/"
      }
    },
    "/events/{event_id}": {
      "get": {
        "summary": "Xem chi tiết một sự kiện trước khi đặt vé",
        "description": "Dùng endpoint này khi người dùng đã chọn hoặc nhắc tới mã sự kiện và muốn xem chi tiết.",
        "operationId": "get_event_detail",
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Mã sự kiện cần xem chi tiết",
              "title": "Event Id"
            },
            "description": "Mã sự kiện cần xem chi tiết"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Event Detail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "x-superapp": {
          "capability": "event.detail.read",
          "sideEffect": "read",
          "riskLevel": "low",
          "requiresConfirmation": false,
          "idempotency": "none"
        },
        "x-risk-level": "low",
        "x-side-effect-type": "read",
        "x-requires-hitl": false,
        "x-idempotency-required": false,
        "x-retry-policy": "safe_retry"
      }
    },
    "/tickets/estimate": {
      "get": {
        "summary": "Ước tính tổng tiền vé trước khi đặt",
        "description": "Dùng endpoint này để báo giá vé dự kiến trước khi người dùng xác nhận đặt vé.",
        "operationId": "estimate_ticket_price",
        "parameters": [
          {
            "name": "event_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Mã sự kiện cần ước tính giá vé",
              "title": "Event Id"
            },
            "description": "Mã sự kiện cần ước tính giá vé"
          },
          {
            "name": "ticket_type",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "standard",
                "vip",
                "student"
              ],
              "type": "string",
              "description": "Loại vé. Giá trị hợp lệ: standard | vip | student",
              "default": "standard",
              "title": "Ticket Type"
            },
            "description": "Loại vé. Giá trị hợp lệ: standard | vip | student"
          },
          {
            "name": "quantity",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "maximum": 10,
              "minimum": 1,
              "description": "Số lượng vé cần đặt",
              "title": "Quantity"
            },
            "description": "Số lượng vé cần đặt"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Estimate Ticket Price"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "x-superapp": {
          "capability": "event.ticket.estimate",
          "sideEffect": "read",
          "riskLevel": "low",
          "requiresConfirmation": false,
          "idempotency": "none"
        },
        "x-risk-level": "low",
        "x-side-effect-type": "read",
        "x-requires-hitl": false,
        "x-idempotency-required": false,
        "x-retry-policy": "safe_retry"
      }
    },
...
    "/bookings/{booking_id}": {
      "get": {
        "summary": "Kiểm tra trạng thái đơn đặt vé",
        "description": "Dùng endpoint này khi người dùng hỏi trạng thái thanh toán hoặc thông tin đơn vé đã đặt.",
        "operationId": "get_booking_status",
        "parameters": [
          {
            "name": "booking_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Mã đơn đặt vé cần kiểm tra",
              "title": "Booking Id"
            },
            "description": "Mã đơn đặt vé cần kiểm tra"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Booking Status"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "x-superapp": {
          "capability": "event.booking.status.read",
          "sideEffect": "read",
          "riskLevel": "low",
          "requiresConfirmation": false,
          "idempotency": "none"
        },
        "x-risk-level": "low",
        "x-side-effect-type": "read",
        "x-requires-hitl": false,
        "x-idempotency-required": false,
        "x-retry-policy": "safe_retry"
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}